REF (string): de-duplicate str_endswith, startswith (#59568)
